isc-dhcpd-server.init.d overhaul
Patches from Peter Marschall to:

isc-dhcp-server.init.d: read pid file from config file

If not already set in /etc/default/isc-dhcp-server, try to read the
pid filename to use from the DHCP deamon's config directive 'pid-file-name'
in the dhcpd.conf file.
Fall back to default /var/run/dhcpd.pid if it is still empty then.
Quote the variable on every use.

Use a variable for /etc/default/isc-dhcp-server in order to have to change
the name only in one place, should the name of this file change.

Allow setting this variable from the calling environment, and fall back to
/etc/default/isc-dhcp-server.

Allow setting the variable DHCPD_CONF in the defaults file
/etc/default/isc-dhcp-server.
Fall back to /etc/dhcp/dhcpd.conf if it is not given there

Use variable OPTIONS in calls to dhcpd.