harmonize logic for setting the host name
Patch from Peter Marschall:

For the udeb case, require the new hostname to be non-empty
in order to set it.

For the non-udeb case, set the hostname in Linux & kFreeBSD only if
1) the new hostname isn't empty
   <=> make sure that we do not reset the hostname to empty

2) the current hostname is empty or '(none)' or 'localhost' or
   is different from the new one from DHCP
   <=> make sure that we set the host name if there is
       no valid one or it changed

This logic is modelled after the one used upstream, and also resembles
more the logic in the dhclient-scripts used in the udeb packages.

This should fix bug #246155