Eliminate sed error message
If no interfaces are supplied to the isc-dhcp-server/interfaces question,
the postinst causes an error from sed:

sed: -e expression #1, char 49: invalid reference \1 on `s' command's RHS

Just don't bother doing any of the substitutions if there's nothing to
substitute it with.

Also quote a variable and eliminate some backticks.