Do a clean between building LDAP and non-LDAP
I had a build failure of the non-LDAP build because it picked up an
LDAP-related .o file that was lying around from the LDAP build.

So do an explicit clean between builds, but don't remove the LDAP-enabled
binary in the clean target.

This feels a bit hacky, but seems like the simplest thing to do for now.